requirements.txt
docopt==0.6.2
google-cloud-bigquery
google-cloud
google-api-core==1.21.0
pandas